+++ changelog.md
@@ -31,6 +31,7 @@ Given a version number MAJOR.MINOR.PATCH:
  - Fixed bug when trying to initialize 'revert to original' when there is no create_from scenario
  - Fixed bug global variable scenario was always set to last scenario.I guess this was introduced in 'Tidy up view.php'
  - Issue 394: bulk measure cost recalculated when reverting a fabric element to original
+ - Issue 351 : Display the cost of an applied measure on line for that element in the tool dialogue
 
 
 ## Version 1.1.0